+++ WargameEngine/WargameEngine/WargameEngine/UI/UITheme.cpp
@@ -90,7 +90,7 @@ void GetFloats(float * array, const char* data, unsigned int max = UINT_MAX)
 	unsigned int i = 0;
 	while (fl != NULL && i < max)
 	{
-		for (unsigned int i = 0; i < strlen(fl); ++i)
+		for (size_t i = 0; i < strlen(fl); ++i)
 		{
 			if (fl[i] == ',') fl[i] = '.';
 		}
